requirements.txt
discord.py
fuzzywuzzy
python-Levenshtein
python-dotenv
aiohttp
pymongo
dnspython